Fix 25 bugs: critical crashes, null safety, memory leaks, logic errors
CRITICAL:
- Fix UUID.fromString('') crash in RoundsEntities bounceBullet/homing/onProjectileHit
- Fix delayed tasks not cancelled in stopGame() - state corruption after stop
- Fix STAT_KEYS.get() null crash in PlayerDataManager.setStat
- Fix GameTeam.valueOf() uncaught exception in BlockListener.onBlockPlace

HIGH:
- Merge bouncePlayer/bouncePl dual fields into single bouncePl
- Fix silence card setting silenceAura instead of silence (block silence was dead code)
- Fix respawn during ROUND_END - now sets invulnerable
- Extract empowerCharge consumption from getEffectiveDamage into consumeEmpowerCharge
- Fix pendingCards memory leak in resetAllCards/clearPendingPicks
- Fix null checks on getAttribute in phantom spawn and hpBoostOnHit
- Fix setHealth without null-guard on attribute in 4 locations

MEDIUM:
- Remove unreachable dead code in checkRoundEnd (aliveTeams.isEmpty())
- Fix getDeadPlayers() returning mutable internal set
- Add null check on loc.getWorld() in BlockListener.saveBlocks
- Add try-catch for distance() in implode damage calculation
- Add null check on getTo() in PlayerMoveEvent handler

LOW:
- Fix shadow parameter in PlayerDataManager.getData(UUID)
- Remove empty applyRoundEffects method body
- Remove double resetAllCards() in debug stop command
- Change roundsToWin/currentRound from double to int
- Fix DebugCommands heal null check on getAttribute
- Add Collections import for unmodifiableSet

+++ src/main/java/com/rounds/game/GameStateManager.java
@@ -61,8 +61,8 @@ public class GameStateManager {
 
         if (state == GameState.WAITING) return null;
 
-        double currentRound = yml.getDouble("current-round", 0);
-        double roundsToWin = yml.getDouble("rounds-to-win", 5);
+        int currentRound = (int) yml.getDouble("current-round", 0);
+        int roundsToWin = (int) yml.getDouble("rounds-to-win", 5);
 
         String loserStr = yml.getString("last-loser");
         GameTeam lastLoser = null;
@@ -94,13 +94,13 @@ public class GameStateManager {
 
     public static class SavedState {
         public final GameState state;
-        public final double currentRound;
-        public final double roundsToWin;
+        public final int currentRound;
+        public final int roundsToWin;
         public final GameTeam lastLoser;
         public final Set<UUID> deadPlayers;
         public final Map<GameTeam, Integer> wins;
 
-        public SavedState(GameState state, double currentRound, double roundsToWin,
+        public SavedState(GameState state, int currentRound, int roundsToWin,
                           GameTeam lastLoser, Set<UUID> deadPlayers, Map<GameTeam, Integer> wins) {
             this.state = state;
             this.currentRound = currentRound;
